feat(vector): make config route read-write
