+++ src/routes/vector/index.ts
@@ -38,7 +38,7 @@ import { vectorServicesApiEndpoint } from './services.ts';
 import { vectorProvidersEndpoint } from './providers.ts';
 import { vectorCostEndpoint } from './cost.ts';
 import { vectorCostsEndpoint } from './costs.ts';
-import { vectorConfigApiEndpoint } from './config-api.ts';
+import { vectorConfigApiEndpoint } from './config.ts';
 import { vectorCollectionsEndpoint } from './collections.ts';
 
 export const vectorRoutes = new Elysia({ prefix: '/api' })
